requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
packaging==24.2 